perf(tui,session): 渲染视口虚拟化 + 有界事件缓冲 + TUI 健壮性修复
- A1: ViewportCache 缓存 flatten 行 + 累计偏移表，render_body 仅重建可见窗口，

  per-frame 成本由 O(transcript) 降到 O(visible_h)；新增 perf 回归测试。

- 事件缓冲 unbounded→bounded (CAPACITY=4096)，backpressure 下 delta 静默丢弃、

  结构事件返回 Full（resume/subagent 同步适配 try_send）。

- B3 清 stale cancel 标志 / B4 Ctrl+C 取消 / poison lock 恢复 /

  C1 panic 结构化 tracing / C4 worker panic→日志文件 / C3 粘贴 256KiB 上限。

- scroll u16→u32，支持 >65535 行超长 transcript。

+++ crates/tui/src/app_task.rs
@@ -43,7 +43,7 @@ pub(crate) async fn switch_session(
     running: &mut bool,
     chat: &mut ChatView,
     history: &mut Vec<String>,
-    scroll: &mut u16,
+    scroll: &mut u32,
     follow: &mut bool,
     sys_tokens: &mut u64,
     queue_items: &mut Vec<(i64, String)>,
